+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.2.3 - 2022-08-01
+
+### Fixed
+
+- Report column 1 in errors instead of 0.
+
+
 ## 0.2.2 - 2022-08-01
 
 ### Fixed